// cinepath-web/src/app/types/request-types.ts
export interface MovieResult {
  poster_path?: string
  adult?: boolean
  overview?: string
  release_date?: string
  genre_ids?: Array<number>
  id?: number
  media_type: 'movie'
  original_title?: string
  original_language?: string
  title?: string
  backdrop_path?: string
  popularity?: number
  vote_count?: number
  video?: boolean
  vote_average?: number
}

export interface TvResult {
  poster_path?: string
  popularity?: number
  id?: number
  overview?: string
  backdrop_path?: string
  vote_average?: number
  media_type: 'tv'
  first_air_date?: string
  origin_country?: Array<string>
  genre_ids?: Array<number>
  original_language?: string
  vote_count?: number
  name?: string
  original_name?: string
}

export interface PersonResult {
  profile_path?: string
  adult?: boolean
  id?: number
  name?: string
  media_type: 'person'
  popularity?: number
  known_for?: Array<MovieResult | TvResult>
}
// ...
export interface PaginatedResponse extends Response {
  page?: number
  total_results?: number
  total_pages?: number
}
// ...
export interface SearchMultiResponse extends PaginatedResponse {
  results?: Array<MovieResult | TvResult | PersonResult>
}